Adding pre-trained model and classical classification algorithms
